apps/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
black==23.10.1
boto3==1.21.35
botocore==1.24.46
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-celery-results==2.5.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filebrowser-safe==1.1.1
flake8==6.1.0
flower==2.0.1
gunicorn==21.2.0
humanize==4.8.0
jmespath==1.0.1
kombu==5.3.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
tzdata==2023.3
urllib3==1.26.18
vine==5.0.0
wcwidth==0.2.9
whitenoise==6.6.0
